jcd_dashboard/requirements.txt
Flask==2.2.5
web3==6.0.0
python-dotenv==1.0.0
apscheduler==3.10.0
Pillow==10.0.0
qrcode==7.4.2
gunicorn==20.1.0